requirements.txt
Flask
flask[async]
python-dotenv
psycopg2-binary
openpyxl
asyncpg
flask-cors
logging